feat(runtime): prove cross-end deterministic dynamic playback
Same frozen v7 package replayed on Native (real winit window + real clock),
WebGL and WebGPU (real Chrome): six rounds, one canonical frame digest
(040c1a9f...). Adds dynamic-frame-v1 canonical contract on both ends,
SceneViewerRoot presentation-frame playback, --smoke-dynamic-package CLI,
replay driver and evidence under test-output/dynamic-runtime-20260919-r1/.

+++ packages/deep-engine-native/src/shadow_pass.rs
@@ -10,7 +10,7 @@ pub fn encode_shadow_cascades(
     culling: &GpuCulling,
     lod: Option<&GpuLod>,
     pipelines: &MeshPipelines,
-    dirty_mask: u8,
+    dirty_mask: u16,
 ) {
     for (cascade_index, layer_view) in shadow_map.layer_views.iter().enumerate() {
         if dirty_mask & (1 << cascade_index) == 0 {
